+++ CHANGELOG.md
@@ -54,7 +54,7 @@
 * Adding `xan (freq|p freq) -X/--approx-algo`.
 * Adding `xan plot -D/--density-gradient, --density-scale, --hide-legend, --hide-x-axis, --hide-y-axis, --hide-all & -Q/--square`.
 * `xan separate` will now avoid emitting columns with an empty name given to `--into`.
-* Adding `xan separate --txt`.
+* Adding `xan separate --txt & --F/--filter`.
 * Adding `pow` & `sqrt` scales.
 
 *Fixes*